requirements.txt
Flask==3.0.3
pymongo==4.17.0
PyJWT==2.12.1
bcrypt==5.0.0
